+++ lexer.h
@@ -34,7 +34,7 @@ extern lexer_t * lexer_new(const char *s);
 extern void      lexer_free(lexer_t *l);
 extern void      lexer_next(lexer_t *l);
 extern int       lexer_string_gen(lexer_t *l, str_ptr s, int tk_ident);
-extern int       lexer_integer(lexer_t *l, long *value);
+extern int       lexer_integer(lexer_t *l, int *value);
 extern int       lexer_number(lexer_t *l, double *value);
 
 #define lexer_string(a, b) lexer_string_gen(a, b, 0)
